Move generateEditorKey in projets.js
